move recent searches into SearchViewController
